v3: Add more details to some errors
